third_party/WebKit/LayoutTests/editing/selection/leak-document-with-selection-inside.html

<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
    <span>This test only runs on DumpRenderTree, as it requires existence of window.internals and cross-domain resource access check disabled.</span>
    <iframe id="f" src='about:blank'></iframe>
    <script>
        description("This test asserts that document doesn't leak when a selection is created inside the document.");

        jsTestIsAsync = true;
        if (!window.internals)
            finishJSTest(); 

        function countNumberOfLiveDocuments() {
            gc();
            return window.internals.numberOfLiveDocuments();
        }

        var numberOfDocumentsBefore = countNumberOfLiveDocuments();
        var numberOfDocumentsAfter = 0;

        var frame = document.getElementById('f');
        frame.onload = function() {
            if (frame.src === 'about:blank') return true;

            // document loaded...

            // create a selection inside iframe
            (function() {
                var contentWindow = frame.contentWindow;
                var element = contentWindow.document.getElementById("t");
                contentWindow.getSelection().setPosition(element, 0);
            })();

            frame.onload = function() {
                // document unloaded...

                numberOfDocumentsAfter = countNumberOfLiveDocuments();
                shouldBe("numberOfDocumentsAfter", "numberOfDocumentsBefore");
                finishJSTest();
            }
            frame.src = 'about:blank';
        }
        frame.src='data:text/html;charset=utf-8,<span id="t"></span>';
    </script>
    <script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
